in progress - added comments
I updated this with some comments since I felt a bit rushed earlier today. This is non-functional in its current state but I guess not bad for only working on it for 2 hours. I think tomorrow or this weekend I'll have a little more time to try and get it working.