Project complete
Since I got the Smurfs Sprint Challenge working today, I decided to model this one exactly after that one, and it works, except for the delete functionality, which is not required. I may come back and see if I can get delete working but for now I'm going to try to relax and not sweat it.